feat(M3): Facility hoan chinh - Branch CRUD (chặn xoa khi con san, tu seed 7 ngay gio mo cua), gio mo cua theo tuan, ngay nghi/le, localize Facilities, 86 tests OK

+++ app/Entities/Payment.php
@@ -28,7 +28,7 @@ class Payment extends Entity
         'invoice_id' => 'integer',
         'payment_code' => 'string',
         'method' => 'string',
-        'amount' => 'decimal',
+        'amount' => 'float',
         'transaction_ref' => 'string',
         'status' => 'string',
         'idempotency_key' => 'string',
